pyproject.toml
[tool.poetry]
name = "dogs-api-drf"
version = "0.1.0"
description = ""
authors = ["PavelSha"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
